Less Bash, more shell portability Signed-off-by: Zdenek Styblik <stybla@turnovfree.net>
This commit is in spirit of Dash and shell portability. In other words, do not
rely on built-in functions of Bash.
One bug found(and fixed in upstream), perhaps some introduced by changes.

Commit addresses what could be called "personal taste" of coding style. It is
not like previous code was wrong or something!